# src/main/java/com/probestack/forgesphere/api/ComplianceScanApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.probestack.forgesphere.api;

import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import com.probestack.forgesphere.model.ComplianceScanDetailsResponse;
import com.probestack.forgesphere.model.ErrorResponse;
import com.probestack.forgesphere.model.SubmitComplianceScanRequest;
import com.probestack.forgesphere.model.SubmitComplianceScanResponse;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-05-13T05:38:50.494838724Z[GMT]")@Validated
@Tag(name = "Compliance Scan", description = "APIs for submitting and fetching compliance scan executions.")
public interface ComplianceScanApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /governance/v1/compliance-scan/{scanId} : Get compliance scan status and result
     * Fetch scan status and results by scanId.
     *
     * @param scanId Unique scan identifier. (required)
     * @return Compliance scan details fetched successfully. (status code 200)
     *         or Scan not found. (status code 404)
     *         or Internal server error. (status code 500)
     */
    @Operation(
        operationId = "getComplianceScanById",
        summary = "Get compliance scan status and result",
        description = "Fetch scan status and results by scanId.",
        tags = { "Compliance Scan" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Compliance scan details fetched successfully.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ComplianceScanDetailsResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Scan not found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/governance/v1/compliance-scan/{scanId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<ComplianceScanDetailsResponse> getComplianceScanById(
        @Parameter(name = "scanId", description = "Unique scan identifier.", required = true, in = ParameterIn.PATH) @PathVariable("scanId") String scanId
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"updatedBy\" : \"updatedBy\", \"scanId\" : \"SCAN-20260511-000001\", \"companyName\" : \"companyName\", \"errorMessage\" : \"errorMessage\", \"rules\" : { \"ruleIds\" : [ \"CR1\", \"CR2\" ], \"ruleTypes\" : [ \"PRE_DEFINED\", \"CUSTOM\" ] }, \"source\" : { \"accessTokenRef\" : \"github-token-secret-ref\", \"bundleName\" : \"Remit2Any.zip\", \"bundleFileId\" : \"file_12345\", \"branch\" : \"main\", \"repositoryUrl\" : \"https://github.com/probestack/payment-service\" }, \"updatedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"scanDate\" : \"2000-01-23T04:56:07.000+00:00\", \"scanResults\" : [ { \"evidence\" : [ { \"filePath\" : \"src/main/java/com/probestack/controller/PaymentController.java\", \"details\" : \"POST /payments endpoint does not have OpenAPI documentation.\", \"lineNumber\" : 42 }, { \"filePath\" : \"src/main/java/com/probestack/controller/PaymentController.java\", \"details\" : \"POST /payments endpoint does not have OpenAPI documentation.\", \"lineNumber\" : 42 } ], \"ruleName\" : \"ruleName\", \"ruleId\" : \"CR1\", \"message\" : \"message\" }, { \"evidence\" : [ { \"filePath\" : \"src/main/java/com/probestack/controller/PaymentController.java\", \"details\" : \"POST /payments endpoint does not have OpenAPI documentation.\", \"lineNumber\" : 42 }, { \"filePath\" : \"src/main/java/com/probestack/controller/PaymentController.java\", \"details\" : \"POST /payments endpoint does not have OpenAPI documentation.\", \"lineNumber\" : 42 } ], \"ruleName\" : \"ruleName\", \"ruleId\" : \"CR1\", \"message\" : \"message\" } ], \"scanOptions\" : { \"saveResult\" : true, \"includeInactiveRules\" : false }, \"processingStartDate\" : \"2000-01-23T04:56:07.000+00:00\", \"createdBy\" : \"createdBy\", \"assetId\" : \"assetId\", \"assetName\" : \"assetName\", \"processingEndDate\" : \"2000-01-23T04:56:07.000+00:00\", \"projectName\" : \"projectName\", \"createDate\" : \"2000-01-23T04:56:07.000+00:00\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * POST /governance/v1/compliance-scan : Submit compliance scan
     * Stores compliance scan request with status submitted and returns response quickly. Actual scan processing is handled asynchronously by ComplianceScanScheduler. 
     *
     * @param submitComplianceScanRequest  (required)
     * @return Compliance scan request submitted successfully. (status code 202)
     *         or Invalid request. (status code 400)
     *         or Internal server error. (status code 500)
     */
    @Operation(
        operationId = "submitComplianceScan",
        summary = "Submit compliance scan",
        description = "Stores compliance scan request with status submitted and returns response quickly. Actual scan processing is handled asynchronously by ComplianceScanScheduler. ",
        tags = { "Compliance Scan" },
        responses = {
            @ApiResponse(responseCode = "202", description = "Compliance scan request submitted successfully.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SubmitComplianceScanResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/governance/v1/compliance-scan",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<SubmitComplianceScanResponse> submitComplianceScan(
        @Parameter(name = "SubmitComplianceScanRequest", description = "", required = true) @Valid @RequestBody SubmitComplianceScanRequest submitComplianceScanRequest
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"scanId\" : \"SCAN-20260511-000001\", \"message\" : \"Compliance scan request submitted successfully.\", \"submittedDate\" : \"2000-01-23T04:56:07.000+00:00\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }

}