# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: |
    API specification for probestack-compliance-service.

    This service manages governance compliance rules and supports compliance scanning
    for microservices, Apigee proxies, and Kong resources.

    ## Microservice

    Microservice Name: probestack-compliance-service

    ## MongoDB Collections

    ### governance_compliance_rules

    Stores compliance rule definitions used by the UI and scan engine.

    Primary Fields:
    - ruleId: Unique rule identifier generated by Java service. Example: CR13
    - ruleName: Compliance rule display name
    - ruleDescription: Detailed rule description
    - ruleType: PRE_DEFINED or CUSTOM
    - ruleOwner: Owner team or person
    - category: DOCUMENTATION, LOGGING, SECURITY, PERFORMANCE, HEALTH, GENERAL
    - severity: REQUIRED, MANDATORY, RECOMMENDED, OPTIONAL
    - enabled: Whether the rule is enabled
    - mandatory: Whether the rule is mandatory
    - displayOrder: UI display order
    - icon: UI icon name
    - status: complaint or non-compliant
    - createDate: Created timestamp
    - createdBy: User who created the rule
    - updatedDate: Last updated timestamp
    - updatedBy: User who last updated the rule

    Recommended Indexes:
    - Unique index on ruleId
    - Compound index on status, category, ruleType

    ### governance_compliance_scans

    Stores submitted, running, completed, and failed compliance scan executions.

    Primary Fields:
...
    - projectName: Project name
...
servers:
...
paths:
...
  /governance/v1/compliance-scan:
    post:
      description: |
        Stores compliance scan request with status submitted and returns response quickly.
        Actual scan processing is handled asynchronously by ComplianceScanScheduler.
      operationId: submitComplianceScan
      requestBody:
        content:
          application/json:
            example:
              projectName: Payment Gateway
              companyName: Probestack
              assetId: PG-001
              assetName: Remit2Any
              assetType: MICROSERVICE
              sourceType: GITHUB
              source:
                repositoryUrl: https://github.com/probestack/payment-service
                branch: main
                accessTokenRef: github-token-secret-ref
              rules:
                ruleTypes:
                - PRE_DEFINED
                - CUSTOM
                ruleIds:
                - CR1
                - CR2
              scanOptions:
                scanMode: FULL
                includeInactiveRules: false
                saveResult: true
              requestedBy: admin@probestack.io
            schema:
              $ref: "#/components/schemas/SubmitComplianceScanRequest"
        required: true
      responses:
        "202":
          content:
            application/json:
              example:
                scanId: SCAN-20260511-000001
                status: submitted
                message: Compliance scan request submitted successfully.
                submittedDate: 2026-05-11T20:00:00Z
              schema:
                $ref: "#/components/schemas/SubmitComplianceScanResponse"
          description: Compliance scan request submitted successfully.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid request.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error.
      summary: Submit compliance scan
      tags:
      - Compliance Scan
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Compliance Scan
  /governance/v1/compliance-scans/run-check:
    post:
      description: Submit compliance scans for one or more selected UI resources. If source.archiveDownloadUrl is provided, that generated archive is scanned before falling back to stored resource source details.
      operationId: runComplianceCheck
      requestBody:
        content:
          application/json:
            example:
              projectName: E-Commerce Platform
              projectId: PRJ-ECOMMERCE
              companyName: ProbeStack
              assetType: MICROSERVICE
              resources:
              - resourceId: RES-MS-AUDIT-PRODUCT-APP
                resourceName: MS-Audit-Product-App
                assetType: MICROSERVICE
                sourceType: BUNDLE_UPLOAD
                source:
                  archiveDownloadUrl: https://storage.probestack.io/codegen-result/archive.zip
              scanOptions:
                scanMode: FULL
                includeInactiveRules: false
                saveResult: true
              requestedBy: admin@probestack.io
            schema:
              $ref: "#/components/schemas/RunComplianceCheckRequest"
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunComplianceCheckResponse"
          description: Compliance run check submitted successfully.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid request.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Matching resources not found.
      summary: Run compliance check for selected resources
      tags:
      - Compliance Scan
  /governance/v1/compliance-scan/{scanId}:
    get:
      description: Fetch scan status and results by scanId.
      operationId: getComplianceScanById
      parameters:
      - description: Unique scan identifier.
        example: SCAN-20260511-000001
        explode: false
        in: path
        name: scanId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceScanDetailsResponse"
          description: Compliance scan details fetched successfully.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Scan not found.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error.
      summary: Get compliance scan status and result
      tags:
      - Compliance Scan
      x-accepts: application/json
      x-tags:
      - tag: Compliance Scan
  /governance/v1/compliance-scans/recent:
    get:
      description: Fetch recent compliance scan executions for the Recent Results UI tab.
      operationId: getRecentComplianceScans
      parameters:
      - description: Project name used by UI context.
        example: E-Commerce Platform
        explode: true
        in: query
        name: projectName
        required: false
        schema:
          type: string
        style: form
      - description: Asset type filter.
        example: MICROSERVICE
        explode: true
        in: query
        name: assetType
        required: false
        schema:
          $ref: "#/components/schemas/AssetType"
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/ComplianceScanDetailsResponse"
                type: array
          description: Recent compliance scans fetched successfully.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error.
      summary: Get recent compliance scan results
      tags:
      - Compliance Scan
components:
  schemas:
    AssetType:
      enum:
      - MICROSERVICE
      - APIGEE
      - KONG
      type: string
    SourceType:
      enum:
      - GITHUB
      - GITLAB
      - BITBUCKET
      - ZIP_UPLOAD
      - BUNDLE_UPLOAD
      type: string
    RuleType:
      enum:
      - PRE_DEFINED
      - CUSTOM
      type: string
    RuleCategory:
      enum:
      - DOCUMENTATION
      - LOGGING
      - SECURITY
      - PERFORMANCE
      - HEALTH
      - GENERAL
      type: string
    RuleSeverity:
      enum:
      - REQUIRED
      - MANDATORY
      - RECOMMENDED
      - OPTIONAL
      type: string
...
    ScanStatus:
      enum:
      - submitted
      - running
      - completed
      - error
      type: string
    ComplianceStatus:
      enum:
      - PENDING
      - COMPLIANT
      - NON_COMPLIANT
      - PARTIAL
      type: string
    ScanMode:
      enum:
      - FULL
      - SELECTED_RULES
      - INCREMENTAL
      type: string
    ScanResultStatus:
      enum:
      - PASSED
      - FAILED
      - WARNING
      - SKIPPED
      type: string
...
    SubmitComplianceScanRequest:
      example:
        scanOptions:
          saveResult: true
          scanMode: null
          includeInactiveRules: false
        requestedBy: admin@probestack.io
        sourceType: null
        assetId: PG-001
        companyName: Probestack
        assetName: Remit2Any
        rules:
          ruleIds:
          - CR1
          - CR2
          ruleTypes:
          - PRE_DEFINED
          - CUSTOM
        source:
          accessTokenRef: github-token-secret-ref
          bundleName: Remit2Any.zip
          bundleFileId: file_12345
          branch: main
          repositoryUrl: https://github.com/probestack/payment-service
        projectName: Payment Gateway
        assetType: null
      properties:
        projectName:
          example: Payment Gateway
          type: string
        companyName:
          example: Probestack
          type: string
        assetId:
          example: PG-001
          type: string
        assetName:
          example: Remit2Any
          type: string
        assetType:
          $ref: "#/components/schemas/AssetType"
        sourceType:
          $ref: "#/components/schemas/SourceType"
        source:
          $ref: "#/components/schemas/ScanSource"
        rules:
          $ref: "#/components/schemas/ScanRules"
        scanOptions:
          $ref: "#/components/schemas/ScanOptions"
        requestedBy:
          example: admin@probestack.io
          type: string
      required:
      - assetId
      - assetName
      - assetType
      - companyName
      - projectName
      - requestedBy
      type: object
    RunComplianceCheckRequest:
      properties:
        assetType:
          $ref: "#/components/schemas/AssetType"
        microserviceId:
          description: Single selected microservice id. Prefer microserviceIds for multi-select.
          example: MS-Audit-Product-App
          type: string
        microserviceIds:
          description: Selected microservice ids from the onboarding API. The compliance service resolves project, name, and archiveDownloadUrl from onboarding.
          items:
            type: string
          type: array
        resources:
          deprecated: true
          description: Deprecated compatibility field. The UI should pass ids only.
          items:
            $ref: "#/components/schemas/RunComplianceResourceRequest"
          type: array
        resourceIds:
          description: Selected Apigee or Kong resource ids from the onboarding API.
          items:
            type: string
          type: array
        resourceNames:
          deprecated: true
          items:
            type: string
          type: array
        sourceType:
          deprecated: true
          $ref: "#/components/schemas/SourceType"
        source:
          deprecated: true
          $ref: "#/components/schemas/ScanSource"
        rules:
          $ref: "#/components/schemas/ScanRules"
        scanOptions:
          $ref: "#/components/schemas/ScanOptions"
        requestedBy:
          example: admin@probestack.io
          type: string
      required:
      - assetType
      type: object
    RunComplianceResourceRequest:
      properties:
        resourceId:
          example: RES-MS-AUDIT-PRODUCT-APP
          type: string
        resourceName:
          example: MS-Audit-Product-App
          type: string
        assetType:
          $ref: "#/components/schemas/AssetType"
        sourceType:
          $ref: "#/components/schemas/SourceType"
        source:
          $ref: "#/components/schemas/ScanSource"
      required:
      - resourceName
      type: object
    RunComplianceCheckResponse:
      properties:
        message:
          example: Compliance run check submitted successfully.
          type: string
        submittedScans:
          example: 2
          format: int32
          type: integer
        scans:
          items:
            $ref: "#/components/schemas/SubmitComplianceScanResponse"
          type: array
      type: object
    ScanSource:
      description: Repository or uploaded bundle details.
      example:
        accessTokenRef: github-token-secret-ref
        bundleName: Remit2Any.zip
        bundleFileId: file_12345
        archiveDownloadUrl: https://storage.probestack.io/codegen-result/archive.zip
        branch: main
        repositoryUrl: https://github.com/probestack/payment-service
      properties:
        repositoryUrl:
          example: https://github.com/probestack/payment-service
          type: string
        branch:
          example: main
          type: string
        accessTokenRef:
          example: github-token-secret-ref
          type: string
        bundleFileId:
          example: file_12345
          type: string
        bundleName:
          example: Remit2Any.zip
          type: string
        archiveDownloadUrl:
          description: Generated bundle archive URL from codegen_result. When provided, the scanner downloads and scans this archive before falling back to Git repository sources.
          example: https://storage.probestack.io/codegen-result/archive.zip
          type: string
      type: object
    ScanRules:
      example:
        ruleIds:
        - CR1
        - CR2
        ruleTypes:
        - PRE_DEFINED
        - CUSTOM
      properties:
        ruleTypes:
          example:
          - PRE_DEFINED
          - CUSTOM
          items:
            $ref: "#/components/schemas/RuleType"
          type: array
        ruleIds:
          example:
          - CR1
          - CR2
          items:
            type: string
          type: array
      required:
      - ruleTypes
      type: object
    ScanOptions:
      example:
        saveResult: true
        scanMode: null
        includeInactiveRules: false
      properties:
        scanMode:
          $ref: "#/components/schemas/ScanMode"
        includeInactiveRules:
          example: false
          type: boolean
        saveResult:
          example: true
          type: boolean
      required:
      - includeInactiveRules
      - saveResult
      - scanMode
      type: object
    SubmitComplianceScanResponse:
      example:
        scanId: SCAN-20260511-000001
        message: Compliance scan request submitted successfully.
        submittedDate: 2000-01-23T04:56:07.000+00:00
        status: null
      properties:
        scanId:
          example: SCAN-20260511-000001
          type: string
        status:
          $ref: "#/components/schemas/ScanStatus"
        message:
          example: Compliance scan request submitted successfully.
          type: string
        submittedDate:
          format: date-time
          type: string
      type: object
    ComplianceScanDetailsResponse:
      example:
        updatedBy: updatedBy
        scanId: SCAN-20260511-000001
        companyName: companyName
        errorMessage: errorMessage
        rules:
          ruleIds:
          - CR1
          - CR2
          ruleTypes:
          - PRE_DEFINED
          - CUSTOM
        source:
          accessTokenRef: github-token-secret-ref
          bundleName: Remit2Any.zip
          bundleFileId: file_12345
          branch: main
          repositoryUrl: https://github.com/probestack/payment-service
        updatedDate: 2000-01-23T04:56:07.000+00:00
        assetType: null
        scanDate: 2000-01-23T04:56:07.000+00:00
        scanResults:
        - severity: null
          result: null
          evidence:
          - filePath: src/main/java/com/probestack/controller/PaymentController.java
            details: POST /payments endpoint does not have OpenAPI documentation.
            lineNumber: 42
          - filePath: src/main/java/com/probestack/controller/PaymentController.java
            details: POST /payments endpoint does not have OpenAPI documentation.
            lineNumber: 42
          ruleType: null
          ruleName: ruleName
          ruleId: CR1
          category: null
          message: message
        - severity: null
          result: null
          evidence:
          - filePath: src/main/java/com/probestack/controller/PaymentController.java
            details: POST /payments endpoint does not have OpenAPI documentation.
            lineNumber: 42
          - filePath: src/main/java/com/probestack/controller/PaymentController.java
            details: POST /payments endpoint does not have OpenAPI documentation.
            lineNumber: 42
          ruleType: null
          ruleName: ruleName
          ruleId: CR1
          category: null
          message: message
        scanOptions:
          saveResult: true
          scanMode: null
          includeInactiveRules: false
        processingStartDate: 2000-01-23T04:56:07.000+00:00
        sourceType: null
        createdBy: createdBy
        assetId: assetId
        compliance: null
        assetName: assetName
        processingEndDate: 2000-01-23T04:56:07.000+00:00
        projectName: projectName
        status: null
        createDate: 2000-01-23T04:56:07.000+00:00
      properties:
        scanId:
          example: SCAN-20260511-000001
          type: string
        projectName:
          type: string
        companyName:
          type: string
        assetId:
          type: string
        assetName:
          type: string
        assetType:
          $ref: "#/components/schemas/AssetType"
        sourceType:
          $ref: "#/components/schemas/SourceType"
        source:
          $ref: "#/components/schemas/ScanSource"
        rules:
          $ref: "#/components/schemas/ScanRules"
        scanOptions:
          $ref: "#/components/schemas/ScanOptions"
        status:
          $ref: "#/components/schemas/ScanStatus"
        compliance:
          $ref: "#/components/schemas/ComplianceStatus"
        scanDate:
          format: date-time
          type: string
        processingStartDate:
          format: date-time
          type: string
        processingEndDate:
          format: date-time
          type: string
        scanResults:
          items:
            $ref: "#/components/schemas/ScanResult"
          type: array
        errorMessage:
          type: string
        createDate:
          format: date-time
          type: string
        createdBy:
          type: string
        updatedDate:
          format: date-time
          type: string
        updatedBy:
          type: string
      type: object
    ScanResult:
      example:
        severity: null
        result: null
        evidence:
        - filePath: src/main/java/com/probestack/controller/PaymentController.java
          details: POST /payments endpoint does not have OpenAPI documentation.
          lineNumber: 42
        - filePath: src/main/java/com/probestack/controller/PaymentController.java
          details: POST /payments endpoint does not have OpenAPI documentation.
          lineNumber: 42
        ruleType: null
        ruleName: ruleName
        ruleId: CR1
        category: null
        message: message
      properties:
        ruleId:
          example: CR1
          type: string
        ruleName:
          type: string
        ruleType:
          $ref: "#/components/schemas/RuleType"
        category:
          $ref: "#/components/schemas/RuleCategory"
        severity:
          $ref: "#/components/schemas/RuleSeverity"
        result:
          $ref: "#/components/schemas/ScanResultStatus"
        message:
          type: string
        evidence:
          items:
            $ref: "#/components/schemas/ScanEvidence"
          type: array
      type: object
    ScanEvidence:
      example:
        filePath: src/main/java/com/probestack/controller/PaymentController.java
        details: POST /payments endpoint does not have OpenAPI documentation.
        lineNumber: 42
      properties:
        filePath:
          example: src/main/java/com/probestack/controller/PaymentController.java
          type: string
        lineNumber:
          example: 42
          type: integer
        details:
          example: POST /payments endpoint does not have OpenAPI documentation.
          type: string
      type: object
...
    ErrorResponse:
      properties:
        timestamp:
          format: date-time
          type: string
        status:
          example: 400
          type: integer
        errorCode:
          example: COMPLIANCE_VALIDATION_ERROR
          type: string
        message:
          example: Invalid request payload.
          type: string
        path:
          example: /governance/v1/compliance-rules
          type: string
        serviceTransactionId:
          example: 3d983a40-fbb6-4df0-b43e-34e1c8f8f4bc
          type: string
      type: object
...